Implement issue 10: Apply changes direct to database.
If you specify an output file, every works exactly as it did before this change.

If you do not specify an output file, dbdeploy will attempt to apply changes to the
database directly, via a jdbc connection.

Files are split using a combination of delimiter and delimitertype (normal or row).
These paraters are intended to work in exactly the same way as those in the
ant sql task: http://ant.apache.org/manual/CoreTasks/sql.html

Default is delimiter=";" and delimitertype="normal".
sqlplus behaviour can be emulated with delimiter="/" delimitertype="row"
osql behaviour can be emulated with delimiter="GO" delimitertype="row"

At the point of this commit, not extensively tested.